add specs delete question

+++ app/controllers/questions_controller.rb
@@ -15,6 +15,7 @@ class QuestionsController < ApplicationController
 
   def create
     @question = Question.new(question_params)
+    @question.author = current_user
     if @question.save
       redirect_to @question, notice: 'Your question successfully created.'
     else
@@ -31,8 +32,12 @@ class QuestionsController < ApplicationController
   end
 
   def destroy
-    question.destroy
-    redirect_to questions_path
+     if current_user.author_of?(question)
+      question.destroy
+      redirect_to questions_path, notice: 'Question was successfully deleted'
+    else
+      redirect_to @question
+    end
   end
 
   private
